@@ -12,27 +12,24 @@
   },
   "devDependencies": {
     "@babel/core": "~7.20.0",
-    "@storybook/addon-actions": "~6.4.18",
-    "@storybook/addon-essentials": "~6.4.18",
-    "@storybook/addon-links": "~6.4.18",
-    "@storybook/cli": "~6.4.18",
-    "@storybook/components": "~6.4.18",
-    "@storybook/core-events": "~6.4.18",
-    "@storybook/react": "~6.4.18",
-    "@storybook/theming": "~6.4.18",
+    "@storybook/cli": "~9.1.6",
+    "@storybook/react": "~9.1.6",
+    "@storybook/react-webpack5": "~9.1.6",
     "@types/heft-jest": "1.0.1",
     "@types/node": "20.17.19",
-    "@types/react-dom": "17.0.25",
-    "@types/react": "17.0.74",
+    "@types/react-dom": "~19.2.3",
+    "@types/react": "~19.2.7",
     "@types/webpack-env": "1.18.8",
     "babel-loader": "~8.2.3",
     "css-loader": "~5.2.7",
     "jest": "~29.3.1",
-    "react-dom": "~17.0.2",
-    "react": "~17.0.2",
+    "react-dom": "~19.2.1",
+    "react": "~19.2.1",
     "style-loader": "~2.0.0",
     "terser-webpack-plugin": "~3.0.8",
     "typescript": "~5.8.2",
-    "webpack": "~4.47.0"
+    "webpack": "~5.98.0",
+    "storybook": "~9.1.6",
+    "@testing-library/dom": "~7.21.4"
   }
 }
