@@ -12,8 +12,8 @@
     "_phase:test": "heft run --only test -- --clean"
   },
   "dependencies": {
-    "react-dom": "~17.0.2",
-    "react": "~17.0.2",
+    "react-dom": "~19.2.1",
+    "react": "~19.2.1",
     "tslib": "~2.8.1"
   },
   "devDependencies": {
@@ -23,13 +23,14 @@
     "@rushstack/heft-lint-plugin": "workspace:*",
     "@rushstack/heft-storybook-plugin": "workspace:*",
     "@rushstack/heft-typescript-plugin": "workspace:*",
-    "@rushstack/heft-webpack4-plugin": "workspace:*",
-    "@rushstack/webpack4-module-minifier-plugin": "workspace:*",
-    "@storybook/react": "~6.4.18",
+    "@rushstack/heft-webpack5-plugin": "workspace:*",
+    "@rushstack/webpack5-module-minifier-plugin": "workspace:*",
+    "@storybook/react-webpack5": "~9.1.6",
+    "@storybook/react": "~9.1.6",
     "@types/heft-jest": "1.0.1",
     "@types/node": "20.17.19",
-    "@types/react": "17.0.74",
-    "@types/react-dom": "17.0.25",
+    "@types/react": "~19.2.7",
+    "@types/react-dom": "~19.2.3",
     "@types/webpack-env": "1.18.8",
     "css-loader": "~5.2.7",
     "eslint": "~9.37.0",
@@ -39,6 +40,9 @@
     "source-map-loader": "~1.1.3",
     "style-loader": "~2.0.0",
     "typescript": "~5.8.2",
-    "webpack": "~4.47.0"
+    "webpack": "~5.98.0",
+    "storybook": "~9.1.6",
+    "@testing-library/dom": "~7.21.4",
+    "@rushstack/module-minifier": "workspace:*"
   }
 }
